Add files via upload
As a student, i'll be creating a static front page of either Myntra or Amazon using only HTML and CSS. This project will help you learn how to structure a webpage, style it to match the chosen site, and make it responsive. It's a great way to practice your web development skills!